requirements/local.txt
# Local development dependencies go here
-r base.txt

# Django extensions
django-extensions==1.6.1
pygraphviz==1.3.1

# Django debug toolbar
django-debug-toolbar==1.4

# Style Enforcement
flake8==2.5.4
isort==4.2.5
pylint-django==0.7.1
pylint-common==0.2.2
pylint-celery==0.3

# Mocking
mock==2.0.0

# Code Coverage
coverage==4.1b2
django_coverage_plugin==1.3

# Test Runner
django-test-plus==1.0.12
pytest-django==2.9.1
pytest-sugar==0.7.1

# Debugging
ipdb==0.9.3

ipython==4.2.0
Werkzeug==0.11.8

# Deploymenyt
Fabric==1.11.1
